+++ oolong/src/commonMain/kotlin/oolong/dispatch/util.kt
@@ -33,5 +33,5 @@ import oolong.Dispatch
  * }
  * ```
  */
-fun <A : Any, B : Any> contramap(dispatch: Dispatch<A>, f: (B) -> A): Dispatch<B> =
+fun <A, B> contramap(dispatch: Dispatch<A>, f: (B) -> A): Dispatch<B> =
     { b -> dispatch(f(b)) }