+++ oolong/src/commonMain/kotlin/oolong/next/util.kt
@@ -47,5 +47,5 @@ import oolong.effect.map
  *     }
  * }
  */
-fun <A : Any, B : Any, C : Any, D : Any> bimap(next: Next<A, B>, fa: (A) -> C, fb: (B) -> D): Next<C, D> =
+fun <A, B, C, D> bimap(next: Next<A, B>, fa: (A) -> C, fb: (B) -> D): Next<C, D> =
     fa(next.first) to map(next.second, fb)