+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `Dispatch` deprecated in favor of `Job`.
 - `oolong.Oolong.runtime` deprecated in favor of `oolong.runtime`.
 - `disposableEffect` deprecated.
+- Allow incoming types to be nullable.
 
 ## [2.0.7] - 2020-08-17
 ### Added